Generalize import combinator pieces to other clauses. (#1308)
Generalize import combinator pieces to other clauses.

For imports and exports, there is logic around how to split and format
the "show" and "hide" combinators. I plan to use the same logic for
"extends", "implements", and "with" clauses in classes so this
generalizes that code to not be specific to imports.

It's also a little simpler and cleaner now.