Format class declarations. (#1309)
Format class declarations.

This handles all of the stuff that can appear outside of the braces in
a class declaration:

- The `class` keyword and any class modifiers.
- The name.
- Type parameters, if any.
- The `extends`, `with`, and/or `implements` clauses.
- The `native` clause.

It doesn't handle members yet.

It also handles ensuring there is a blank line before and after class
declarations. In the process of doing that, I also migrated over similar
code for function declarations where we ensure there is a blank line
after non-empty functions.