Get rid of the Score class and inline it into Solution.
It was an abstraction that didn't carry its weight: the only place that
used it was Solution, and in only one place.

After this change, I intend to merge some of the state in CodeWriter
into Solution so that CodeWriter mutates the Solution as it runs instead
of copying the results of writing into a new immutable Solution.

Immutable solutions are nice, but an optimization I'm planning is
simpler to implement if we can write some state into the Solution
(which pieces are in overflowing lines) during the code writing process.