Format mixin application classes. (#1310)
Format mixin application classes.

This mostly reuses the existing createType() code with a couple of
tweaks since mixin application classes don't have bodies and have the
"= Superclass" part before the "with" clause.

You'll notice that the modifiers test case here takes a couple of
seconds to run. That's because the new back end's line splitter doesn't
do any memoization or divide and conquering, so it's brute-forcing the
whole compilation unit to figure out which of the mixin applications
need splitting.

I don't think it's intolerable yet as far as how long it takes the tests
to run, but this means it's probably a good time for me to start
looking into optimization soon.