Format for-in loops.
I ended up refactoring the code for handling the for parts and inlining
it all into visitForStatement() because it turns out that there isn't
as much sharing between C-style for loops and for-in loops as I was
expecting.

Thanks to the magic of pattern matching and shared case bodies, I was
still able to get it all into a single flattened switch statement for
all the kinds of for parts.