Less branchy expanding (#1311)
* Get rid of the Score class and inline it into Solution.

It was an abstraction that didn't carry its weight: the only place that
used it was Solution, and in only one place.

After this change, I intend to merge some of the state in CodeWriter
into Solution so that CodeWriter mutates the Solution as it runs instead
of copying the results of writing into a new immutable Solution.

Immutable solutions are nice, but an optimization I'm planning is
simpler to implement if we can write some state into the Solution
(which pieces are in overflowing lines) during the code writing process.

* Get rid of separate indent and nesting fields.

As far as I can tell (unlike in the old formatter), it isn't doing
anything useful. So, for now at least, removing it to simplify things
a little.

* When expanding, only use the first unsolved piece in the first bad line.

The line splitting solver works by incrementally binding one piece to
a state at a time, expanding out a tree of partial solutions until it
finds one that fits.

Currently, the solver uses no logic to choose which pieces to bind to
states when expanding. It just brute forces them all from beginning to
end. This means it spends a bunch of time adding line splits to places
that don't need them and generating losing solutions that later get
discarded.

Instead, with this change, we take the same optimization that the old
line splitter uses: When expanding a solution, we only look at the first
solvable piece in the first problematic line.

A "solvable" piece is one that allows multiple states and hasn't already
been bound in the current solution. A problematic line is one that
overflows the page width or contains an invalid newline.

This optimization lets the solver skip past all of the non-problematic
pieces that are in regions of code that already fit.

There will probably be more optimization work to do later, but this
gets the tests back down to <1 second.