Format fields and methods. (#1313)
Format fields and methods.

Almost all of the real work was already done to handle variables and
functions, so this just extends that to fields and methods defined in
classes and covers a few loose ends:

- External declarations in their various forms.
- The covariant modifier on parameters and fields.
- Operator declarations (which apparently were never tested on the old
  style!).
- Getters and setters.
- The `static`, `abstract`, and `late` modifiers.

I also tweaked the cost heuristics around splitting between return
types, parameter lists, and `=>` expression bodies. It looks pretty
weird to split after a return type, so I made that more costly than
splitting at the body or parameter list.

I also decided to put the tests for member declarations inside
declaration/ instead of member/. My impression is that there won't be
a huge number of separate files in declaration/ and member/, so it's
easier to just merge them together.