Format mixin declarations.
Almost nothing interesting here. The clauses are all handled the same
way we handle clauses in a class declaration. We treat the `on` clause
like the `extends` clause in a class, which means the other clauses can
split while allowing the on clause to stay on the top line if it fits.

Mixin classes (and their various modifiers) were already handled because
those are class declarations with a mixin modifier. Now there are tests
for them.