Format enum declarations without members.
When an enum declaration contains only constants, it's pretty similar
to a comma-delimited list, and we format it much the same. The only
wrinkle is that an enum declaration may have a trailing semicolon but
no members after it. In that case, we discard the `;` and treat it like
a normal enum declaration.

That requires a little extra care to ensure that we don't drop comments
preceding the `;` on the floor. Added a lot of tests to cover this.

Also, it means that we treat `;` like whitespace, so also loosened
equalIgnoringWhitespace() to ignore semicolons too.