Format enum declarations with members.
Since an enum with members has a required `;` between the constants and
members, we discard any redundant trailing comma after the last constant
but before the `;`.

Whenever we discard tokens, we have to worry about losing comments, so
this tests all of the edge cases there. In practice, most of these
tests are pathological, and the output doesn't matter that much as long
as all of the comments are preserved.