Ensure that line comments always get newlines after them.
The formatter currently mostly relies on the surrounding piece's
setAllowedWhitespace() rules to make sure that a line comment forces
the surrounding piece to split in a way that ensures there is a newline
after the comment.

This works correctly in lots of places where line comments are actually
expected, but isn't reliable when a line comment occurs in a weird
location. When that happens, it's important that we still always put a
newline after the comment. Otherwise the code is meaningfully changed.

This does that. Whenever a TextPiece ends in a line comment, it will
know that it has a trailing newline. It then always writes that to the
CodeWriter. That leads to redundant newlines in cases where the
surrounding piece is already going to put a newline, so I also moved
some of the "pending" newline collapsing code from PieceWriter farther
down the pipeline to CodeWriter.

This doesn't change the behavior of any of the current tests. It will
fix many untested corners of the language. I'm working on some new tests
for comments in if statements that will be fixed by this change, but
I'm also changing the actual formatting of if statements, so I want to
do that in a separate change.