Improve how parenthesized expressions in control flow are handled.
In the process of writing tests for if statements, I realized that the
way if conditions are formatted was different than switch values. And
it turns out that, based on looking at how all of the control flow in
Flutter is hand-formatted, the way I was handling switches was more
complex than needed.

It seems like the right behavior we want is to just visit the expression
directly and let the parentheses attach directly to it.

So I updated if, while, and switch to all do that.

Also added tests for comments in if statements.