Format if elements.
I was hoping to be able to reuse more of the createIf() code in
PieceFactory for both if elements and if statements, but there are
enough differences between the two that that didn't work out, so I just
inlined the code for both of them directly in AstNodeVisitor.

I was able to reuse the existing IfPiece, though, which is nice.