Format for elements.
This mostly uses the existing code for formatting the `for (...)` part
of for statements, which I moved into PieceFactory so that it can be
reused.

In the process of adding tests for this, I noticed that a newline in
the initializer of a for-in loop forced the `in` to split, which looks
weird, so I fixed that too, for both for statements and for elements.