Change code blocks to use indentation instead of backticks.
Not exactly the most impactful change. :) But in a previous PR, Nate
suggested that old-style Markdown indentation for code blocks would be
better than backticks since it's shorter and we aren't actually running
the comments through a Markdown formatter anyway.

This converts all of them over in on Go so the codebase is consistent.