Format cascades.
There were a couple of tricky corners, but this wasn't as bad as I'd
feared. Most of ChainBuilder worked great for it.

In the process of doing this and migrating the tests, I noticed another
kind of expression that should allow block formatting in an argument
list: function expression invocations.

That's a `(...)` where the thing on the left is some kind of expression
and not simply a name, like: `(foo + bar)(args)`.

Added support for that and added some tests of block formatting for
those too.