Format super expressions. (#1366)
Apparently the old formatter has *zero* tests of this. I went ahead and
wrote tests of all of the ways `super` appears in the grammar. That's
probably overkill given that the analyzer models `super` as a standalone
expression, but that's not actually how it works in the language itself.
(In Dart, `print(super)` isn't syntactically valid.)

I figured it made sense to write tests closer to the grammar.