Refactor PieceStateSet, Solution, and CodeWriter.
I initially wanted Solution to be completely immutable. That meant it
needed to know the full cost and text of the solution at construction
time. But to calculate that, CodeWriter needs to know the set of bound
states. So I split that out into a separate PieceStateClass.

Also, the cost and overflow are calculated incrementally, so I have
CodeWriter its own mutable copy of that state and other stuff.

The result was that we created a PieceStateSet, passed that to a
CodeWriter, which calculated everything and then created a Solution.

It worked OK, but there's a lot of duplicate state between Solution and
CodeWriter. I'm working on some optimizations that will lead to the
bound states being mutable, so I figured it was best to just accept that
Solution is itself going to be built up mutably (but won't be mutated
once CodeWriter is done with it).

This commit just does that reorganization. There are no behavioral
changes.