Prune solutions derived from invalid ones.
A solution is invalid if a child piece writes a newline at a point in
time where a parent disallows newlines.

When that happens, the parent's may be explicitly bound to a state by
the solution, or it may still be in an unbound state (and thus using
the default). If unbound, then even though this solution is invalid, it
may lead to later valid ones by binding that parent piece to a different
state.

But once a solution binds a piece to a state, all solutions derived
from that state keep the same binding. So if we hit an invalid solution
and the piece whose newline constraint was already violated, we can
discard that solution and all possible solutions derived from it.

This skips an entire potentially huge subtree from the solution space
being explored.

In an example giant deeply nested function call from the benchmark,
before this optimization, the solver tries 329,057 solutions before
finding a winning one. With this optimization, it only takes 5,405.