Reorganize how Solution handles pinned states and constraints.
In the process of turning ListElement into a Piece, I ran into a couple
of bugs in how pinned states and constraints between pieces are handled.

The tests didn't catch those bugs because there are basically two bugs
that cancel out. The IfPiece was incorrectly pinning an empty else block
in an if statement to split. But the solver was incorrectly ignoring
pinned states in some cases. This fixes both bugs.

With this change, when generating a Solution, we eagerly bind as many
Pieces to States as we can when the Solution is first created before
running CodeWriter. To do that, we find all of the pinned Pieces and
bind them to their pinned States, then propagate any constraints from
that to other child Pieces.

This fixes a couple of bugs, but also, I think, leads to a cleaner
module where at Solution creation time, we fill in as many piece states
as we can based on what we know statically about the Piece tree (its
pinned states and their constraints). Then during solution exploring and
line splitting, we don't have to worry about pinned states at all.

This will, I think, align with future optimizations where we do some
other piece pinning before running the solver.