Turn ListElement into a Piece.
As with the previous commit, we should be able to format arguments and
collection literal elements separately when the surrounding argument
list or collection is split. But that requires each ListElement to be
an actual Piece object.

This commit does that.

It's trickier for ListElement than SequenceElement was because whether
or not a ListElement needs a comma can depend how the surrounding
ListPiece splits. To model that, I gave ListElementPiece two states,
one with and one without the comma. Then I use pinning and constraints
in ListPiece to switch on commas for the correct elements.