Format infix operands separately.
We can't apply this for the first and last operand since there may be
code on the line before the first one or after the last. But this commit
applies it to all interior operands when the InfixPiece has more than
two.

Hand-authored that has very large infix operator chains isn't very
common, but it comes up pretty frequently in generated code.

Because the first and last operands aren't formatted separately, this
doesn't completely solve the performance of infix operators. A future
optimization should help more.