Solve subtrees separately (#1376)
Support formatting Piece subtrees separately from the surrounding Piece.

This is the biggest optimization hammer because it takes a problem that
is combinatorial in the total number of Pieces in the tree and breaks
the Piece tree into separate subtrees that can each be formatted
independently.

Further, it allows reusing subtree Solutions across different Solutions
for the surrounding Piece trees.

However, knowing when this optimization can be applied is tricky. The
whole reason that solving is combinatorial is because in many regions of
code, we want to find the globally optimal solution and Pieces are
interdependent since how one Piece formats determines where in the
current line the next Piece starts, how indented it is, possibly other
constraints, etc.

I tried but wasn't able to find a robust automatic way for CodeWriter
to tell when formatting a given child Piece can be done independently.
Instead, we rely on each Piece to tell CodeWriter that a child can be
hoisted out by passing a `separate` flag. In theory, you want this to be
`true` as often as possible. But passing `true` when the child isn't
actually separate can lead to incorrect formatting results.

In practice, the basic rule of thumb is a child can usually be formatted
separately if the parent piece knows that there is a newline before it,
there will be a newline after it, and there are no constraints being
applied to any of its children. Further, we mostly only need to care
about this optimization for the small number of pieces that are commonly
used and can contain an unbounded number of children. That's:

- SequencePiece
- ChainPiece
- ListPiece
- InfixPiece

In order to format the children of those separately, their children need
to actually be pieces, so this also turns SequenceElement and
ListElement into Piece classes.

Also added a handful of benchmarks to validate this optimization on
those various kinds of pieces.