Pin Pieces if we can determine they will split just by their children.
If, for example, the elements inside a ListPiece are longer than the
page width when you add up all of their characters, then the ListPiece
will definitely split. But the Solver will try every possible way to
*not* split that Piece and waste a ton of time doing so.

In particular, since the Solver reaches outermost Pieces first, it
will fully explore the entire Solution tree where those don't split
before finally giving up and letting them split. But the outmost Pieces
also have the largest number of Solutions that can be derived from them,
so the Solver gets stuck exploring entire huge pointless branches.

That's unnecessary because in some cases it's obvious that a Piece will
split just because of how big it's contents are. Assuming an
80-character width, a ListPiece for a collection literal is *definitely*
going to split if it's contents are >80 characters long. We don't need
the Solver to figure that out.

This PR adds an optimization where before solving, we ask each Piece if
it can be pinned to a state based just on the length of its children
and whether there are any newlines in them.

Then, for ListPiece and InfixPiece, we implement that optimization.

Also, added a bunch of benchmarks that validate this optimization for
various syntax using those pieces.

With this optimization, the new formatter is now only about 30% slower
than the old one on the large benchmark. (It's faster than the old
formatter and slower for some of the other various benchmarks.) That's
at the point where we can start looking for constant-time optimizations.

\o/